requirements.txt
fastapi
uvicorn
aiogram
telethon
pandas
python-dotenv
